Create package, LOST+FOUND
