+++ app/src/main/java/com/rymmmmm/hook/CustomSplash.java
@@ -1,4 +1,4 @@
-package nil.nadph.qnotified.hook.rikka;
+package com.rymmmmm.hook;
 
 import android.content.res.AssetManager;
