+++ app/src/main/java/com/rymmmmm/hook/DisableAvatarDecoration.java
@@ -1,10 +1,10 @@
-package nil.nadph.qnotified.hook.rikka;
+package com.rymmmmm.hook;
 
 import android.os.Looper;
+import android.view.View;
 import android.widget.Toast;
 
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 
 import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedBridge;
@@ -20,15 +20,17 @@ import static nil.nadph.qnotified.util.Utils.TOAST_TYPE_ERROR;
 import static nil.nadph.qnotified.util.Utils.getApplication;
 import static nil.nadph.qnotified.util.Utils.log;
 
-//屏蔽戳一戳动画
-public class DisablePokeEffect extends BaseDelayableHook {
-    public static final String rq_disable_poke_effect = "rq_disable_poke_effect";
-    private static final DisablePokeEffect self = new DisablePokeEffect();
+//屏蔽头像挂件
+public class DisableAvatarDecoration extends BaseDelayableHook {
+    public static final String rq_disable_avatar_decoration = "rq_disable_avatar_decoration";
+    private static final DisableAvatarDecoration self = new DisableAvatarDecoration();
     private boolean isInit = false;
 
-    public static DisablePokeEffect get() {
+
+    public static DisableAvatarDecoration get() {
         return self;
     }
+
 
     @Override
     public int getEffectiveProc() {
@@ -44,36 +46,29 @@ public class DisablePokeEffect extends BaseDelayableHook {
     public boolean init() {
         if (isInit) return true;
         try {
-            for (Method m : Initiator._GivingHeartItemBuilder().getDeclaredMethods()) {
-                Class<?>[] argt = m.getParameterTypes();
-                if (m.getName().equals("a") && argt.length == 3 && !Modifier.isStatic(m.getModifiers())) {
+            for (Method m : Initiator.load("com.tencent.mobileqq.vas.PendantInfo").getDeclaredMethods()) {
+                if (m.getReturnType() == void.class) {
+                    Class<?>[] argt = m.getParameterTypes();
+                    if (argt.length != 5) continue;
+                    if (argt[0] != View.class) continue;
+                    if (argt[1] != int.class) continue;
+                    if (argt[2] != long.class) continue;
+                    if (argt[3] != String.class) continue;
+                    if (argt[4] != int.class) continue;
                     XposedBridge.hookMethod(m, new XC_MethodHook() {
                         @Override
                         protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
                             if (LicenseStatus.sDisableCommonHooks) return;
                             if (!isEnabled()) return;
-                            // param.setResult(null);// 此处不应为null
-                            if (param.getResult().getClass().isPrimitive()) {// 判断是boolean (基本类型)
-                                param.setResult(false);
-                            }
+                            param.setResult(null);
                         }
                     });
                 }
-//                "fangdazhao" need to fix.
-//                if (m.getName().equals("b") && m.getParameterTypes().length == 2 && !Modifier.isStatic(m.getModifiers())) {
-//                    XposedBridge.hookMethod(m, new XC_MethodHook() {
-//                        @Override
-//                        protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
-//                            if (LicenseStatus.sDisableCommonHooks) return;
-//                            if (!isEnabled()) return;
-//                            param.setResult(null);
-//                        }
-//                    });
-//                }
             }
             isInit = true;
             return true;
-        } catch (Throwable e) {
+        } catch (Throwable t) {
+            Utils.log(t);
             return false;
         }
     }
@@ -84,10 +79,20 @@ public class DisablePokeEffect extends BaseDelayableHook {
     }
 
     @Override
+    public boolean isEnabled() {
+        try {
+            return ConfigManager.getDefaultConfig().getBooleanOrFalse(rq_disable_avatar_decoration);
+        } catch (Exception e) {
+            log(e);
+            return false;
+        }
+    }
+
+    @Override
     public void setEnabled(boolean enabled) {
         try {
             ConfigManager mgr = ConfigManager.getDefaultConfig();
-            mgr.getAllConfig().put(rq_disable_poke_effect, enabled);
+            mgr.getAllConfig().put(rq_disable_avatar_decoration, enabled);
             mgr.save();
         } catch (final Exception e) {
             Utils.log(e);
@@ -101,16 +106,6 @@ public class DisablePokeEffect extends BaseDelayableHook {
                     }
                 });
             }
-        }
-    }
-
-    @Override
-    public boolean isEnabled() {
-        try {
-            return ConfigManager.getDefaultConfig().getBooleanOrFalse(rq_disable_poke_effect);
-        } catch (Exception e) {
-            log(e);
-            return false;
         }
     }
 }
