+++ app/src/main/java/com/rymmmmm/hook/DisableColorNickName.java
@@ -1,4 +1,4 @@
-package nil.nadph.qnotified.hook.rikka;
+package com.rymmmmm.hook;
 
 import android.os.Looper;
 import android.widget.Toast;
