+++ app/src/main/java/com/rymmmmm/hook/DisableShakeWindow.java
@@ -1,4 +1,4 @@
-package nil.nadph.qnotified.hook.rikka;
+package com.rymmmmm.hook;
 
 import android.os.Looper;
 import android.widget.Toast;
@@ -20,16 +20,15 @@ import static nil.nadph.qnotified.util.Utils.TOAST_TYPE_ERROR;
 import static nil.nadph.qnotified.util.Utils.getApplication;
 import static nil.nadph.qnotified.util.Utils.log;
 
-//屏蔽掉落小表情
-public class DisableDropSticker extends BaseDelayableHook {
-    public static final String rq_disable_drop_sticker = "rq_disable_drop_sticker";
-    private static final DisableDropSticker self = new DisableDropSticker();
+//屏蔽抖动窗口 作用暂时不明
+public class DisableShakeWindow extends BaseDelayableHook {
+    public final static String rq_disable_shake_window = "rq_disable_shake_window";
+    private final static DisableShakeWindow self = new DisableShakeWindow();
     private boolean isInit = false;
 
-    public static DisableDropSticker get() {
+    public static DisableShakeWindow get() {
         return self;
     }
-
 
     @Override
     public int getEffectiveProc() {
@@ -45,25 +44,27 @@ public class DisableDropSticker extends BaseDelayableHook {
     public boolean init() {
         if (isInit) return true;
         try {
-            for (Method m : Initiator._ConfigHandler().getDeclaredMethods()) {
-                Class<?>[] argt = m.getParameterTypes();
-                if (m.getName().equals("f") && !Modifier.isStatic(m.getModifiers()) && argt.length == 1) {
-                    XposedBridge.hookMethod(m, new XC_MethodHook() {
-                        @Override
-                        protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
-                            if (LicenseStatus.sDisableCommonHooks) return;
-                            if (!isEnabled()) return;
-                            param.setResult(null);
-                        }
-                    });
+            for (int i = 1; i < 4; i++) {
+                for (Method m : Initiator.load("com.tencent.mobileqq.activity.aio.helper.AIOShakeHelper$" + i).getDeclaredMethods()) {
+                    if (m.getName().equals("run") && !Modifier.isStatic(m.getModifiers())) {
+                        XposedBridge.hookMethod(m, new XC_MethodHook() {
+                            @Override
+                            protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
+                                if (LicenseStatus.sDisableCommonHooks) return;
+                                if (!isEnabled()) return;
+                                param.setResult(null);
+                            }
+                        });
+                    }
                 }
             }
             isInit = true;
             return true;
-        } catch (Exception e) {
+        } catch (Throwable e) {
             Utils.log(e);
             return false;
         }
+
     }
 
     @Override
@@ -75,9 +76,9 @@ public class DisableDropSticker extends BaseDelayableHook {
     public void setEnabled(boolean enabled) {
         try {
             ConfigManager mgr = ConfigManager.getDefaultConfig();
-            mgr.getAllConfig().put(rq_disable_drop_sticker, enabled);
+            mgr.getAllConfig().put(rq_disable_shake_window, enabled);
             mgr.save();
-        } catch (final Throwable e) {
+        } catch (final Exception e) {
             Utils.log(e);
             if (Looper.myLooper() == Looper.getMainLooper()) {
                 Utils.showToast(getApplication(), TOAST_TYPE_ERROR, e + "", Toast.LENGTH_SHORT);
@@ -95,7 +96,7 @@ public class DisableDropSticker extends BaseDelayableHook {
     @Override
     public boolean isEnabled() {
         try {
-            return ConfigManager.getDefaultConfig().getBooleanOrFalse(rq_disable_drop_sticker);
+            return ConfigManager.getDefaultConfig().getBooleanOrFalse(rq_disable_shake_window);
         } catch (Exception e) {
             log(e);
             return false;
