+++ app/src/main/java/nil/nadph/qnotified/dialog/RikkaCustomDeviceModelDialog.java
@@ -17,7 +17,7 @@ import java.io.IOException;
 
 import nil.nadph.qnotified.R;
 import nil.nadph.qnotified.config.ConfigManager;
-import nil.nadph.qnotified.hook.rikka.CustomMsgTimeFormat;
+import com.rymmmmm.hook.CustomMsgTimeFormat;
 import nil.nadph.qnotified.ui.CustomDialog;
 import nil.nadph.qnotified.util.NonNull;
 import nil.nadph.qnotified.util.Nullable;
