+++ app/src/main/java/nil/nadph/qnotified/dialog/RikkaCustomMsgTimeFormatDialog.java
@@ -19,7 +19,7 @@ import java.util.Date;
 
 import nil.nadph.qnotified.R;
 import nil.nadph.qnotified.config.ConfigManager;
-import nil.nadph.qnotified.hook.rikka.CustomMsgTimeFormat;
+import com.rymmmmm.hook.CustomMsgTimeFormat;
 import nil.nadph.qnotified.ui.CustomDialog;
 import nil.nadph.qnotified.util.NonNull;
 import nil.nadph.qnotified.util.Nullable;
