+++ app/src/main/java/nil/nadph/qnotified/dialog/RikkaCustomSplash.java
@@ -18,9 +18,8 @@ import java.io.File;
 import java.io.IOException;
 
 import nil.nadph.qnotified.R;
-import nil.nadph.qnotified.activity.SettingsActivity;
 import nil.nadph.qnotified.config.ConfigManager;
-import nil.nadph.qnotified.hook.rikka.CustomMsgTimeFormat;
+import com.rymmmmm.hook.CustomMsgTimeFormat;
 import nil.nadph.qnotified.ui.CustomDialog;
 import nil.nadph.qnotified.util.NonNull;
 import nil.nadph.qnotified.util.Nullable;
