+++ app/src/main/java/nil/nadph/qnotified/dialog/RikkaDialog.java
@@ -20,9 +20,9 @@ import androidx.core.view.ViewCompat;
 
 import nil.nadph.qnotified.activity.IphoneTitleBarActivityCompat;
 import nil.nadph.qnotified.hook.BaseDelayableHook;
-import nil.nadph.qnotified.hook.rikka.OneTapTwentyLikes;
-import nil.nadph.qnotified.hook.rikka.RemoveSendGiftAd;
-import nil.nadph.qnotified.hook.rikka.ShowMsgCount;
+import com.rymmmmm.hook.OneTapTwentyLikes;
+import com.rymmmmm.hook.RemoveSendGiftAd;
+import com.rymmmmm.hook.ShowMsgCount;
 import nil.nadph.qnotified.ui.DummyDrawable;
 import nil.nadph.qnotified.ui.ResUtils;
 import nil.nadph.qnotified.ui.ViewBuilder;
