+++ app/src/main/java/nil/nadph/qnotified/hook/AbsDelayableHook.java
@@ -52,21 +52,21 @@ import me.singleneuron.hook.decorator.SimpleCheckIn;
 import me.singleneuron.qn_kernel.dispacher.ItemBuilderFactoryHook;
 import me.singleneuron.qn_kernel.dispacher.StartActivityHook;
 import nil.nadph.qnotified.config.SwitchConfigItem;
-import nil.nadph.qnotified.hook.rikka.BaseApk;
-import nil.nadph.qnotified.hook.rikka.CustomDeviceModel;
-import nil.nadph.qnotified.hook.rikka.CustomMsgTimeFormat;
-import nil.nadph.qnotified.hook.rikka.CustomSplash;
-import nil.nadph.qnotified.hook.rikka.DefaultFont;
-import nil.nadph.qnotified.hook.rikka.DisableAvatarDecoration;
-import nil.nadph.qnotified.hook.rikka.DisableColorNickName;
-import nil.nadph.qnotified.hook.rikka.DisableDropSticker;
-import nil.nadph.qnotified.hook.rikka.DisableEnterEffect;
-import nil.nadph.qnotified.hook.rikka.DisablePokeEffect;
-import nil.nadph.qnotified.hook.rikka.IgnoreDiyCard;
-import nil.nadph.qnotified.hook.rikka.OneTapTwentyLikes;
-import nil.nadph.qnotified.hook.rikka.RemoveMiniProgramAd;
-import nil.nadph.qnotified.hook.rikka.RemoveSendGiftAd;
-import nil.nadph.qnotified.hook.rikka.ShowMsgCount;
+import com.rymmmmm.hook.BaseApk;
+import com.rymmmmm.hook.CustomDeviceModel;
+import com.rymmmmm.hook.CustomMsgTimeFormat;
+import com.rymmmmm.hook.CustomSplash;
+import com.rymmmmm.hook.DefaultFont;
+import com.rymmmmm.hook.DisableAvatarDecoration;
+import com.rymmmmm.hook.DisableColorNickName;
+import com.rymmmmm.hook.DisableDropSticker;
+import com.rymmmmm.hook.DisableEnterEffect;
+import com.rymmmmm.hook.DisablePokeEffect;
+import com.rymmmmm.hook.IgnoreDiyCard;
+import com.rymmmmm.hook.OneTapTwentyLikes;
+import com.rymmmmm.hook.RemoveMiniProgramAd;
+import com.rymmmmm.hook.RemoveSendGiftAd;
+import com.rymmmmm.hook.ShowMsgCount;
 import nil.nadph.qnotified.step.Step;
 import nil.nadph.qnotified.util.Utils;
