+++ app/src/main/java/nil/nadph/qnotified/lifecycle/ActProxyMgr.java
@@ -91,7 +91,8 @@ public class ActProxyMgr {
         return className.startsWith("nil.nadph.qnotified.")
             || className.startsWith("me.zpp0196.qqpurify.activity.")
             || className.startsWith("me.singleneuron.")
-            || className.startsWith("me.ketal.activity");
+            || className.startsWith("me.ketal.activity.")
+            || className.startsWith("com.rymmmmm.activity.");
     }
 
     public static boolean isResourceInjectionRequired(@NonNull String className) {
